new settings page combines all settings on one page
