+++ htdocs/templates2/ocstyle/mydetails.tpl
@@ -6,97 +6,74 @@
 {* OCSTYLE *}
 <div class="content2-pagetitle">
 	<img src="resource2/{$opt.template.style}/images/profile/32x32-profile.png" style="margin-right: 10px;" width="32" height="32" alt="" />
-	{t}My profile details{/t}
+	{t}My public profile settings{/t} <!-- TODO: Translation -->
 </div>
+
+{include file="settingsmenu.tpl"}
 
 <form action="mydetails.php" method="post" style="display:inline;">
 	<input type="hidden" name="action" value="change" />
-
-<table class="table">
-	<tr>
-		<td colspan="3">
-			<span class="boldtext">{t}The following additional information is stored in your userprofile:{/t}</span>
-		</td>
-	</tr>
-
-	{if $error==true || $errorlen==true}
+	<table class="table">
 		<tr>
-			<td class="errormsg" colspan="3">
-				{t}Error while saving.{/t}<br />
-				{if $error==true}{t}Illegal characters found in{/t}{$errormsg|escape}<br />{/if}
-				{if $errorlen==true}{t}Field values too long in{/t}{$errormsglen|escape}<br />{/if}
-				{t}Original values were restored.{/t}
+			<td colspan="3">
+				<span class="boldtext">{t}The following additional information is shown in your <a href="viewprofile.php">public profile</a>:{/t}</span><br /> <!-- TODO: Translation -->
+				<img src="resource2/{$opt.template.style}/images/misc/hint.gif" border="0" width="15" height="11" alt="" align="middle" />
+				<span class="systemlink" style="font-size:10px;">{t}Only the <span class="public-setting">green entries</span> are visible in your <a href="viewprofile.php">public profile</a>.{/t}</span>
 			</td>
 		</tr>
-	{/if}
 
-	<tr><td class="spacer" colspan="3"></td></tr>
+		{if $error==true || $errorlen==true}
+			<tr>
+				<td class="errormsg" colspan="3">
+					{t}Error while saving.{/t}<br />
+					{if $error==true}{t}Illegal characters found in{/t}{$errormsg|escape}<br />{/if}
+					{if $errorlen==true}{t}Field values too long in{/t}{$errormsglen|escape}<br />{/if}
+					{t}Original values were restored.{/t}
+				</td>
+			</tr>
+		{/if}
+		<tr><td class="spacer" colspan="3"></td></tr>
 
-	{foreach from=$useroptions item=useropt}
+		{include file="displayuseroptions.tpl" dUseroptions=$useroptions1}
+
 		<tr>
-			<td style="vertical-align:top; width:10px"><nobr>{$useropt.name|escape}{t}#colonspace#{/t}:</nobr></td>
+			<td valign=top>{t}Use PMR:{/t}</td>  <!-- TODO: Translation -->
 			<td>
 				{if $edit==true}
-					{if $useropt.option_input=="text"}
-						<input type="text" name="inp{$useropt.id}" value="{$useropt.option_value|escape}" class="input200" />
-					{/if}
-					{if $useropt.option_input=="textarea"}
-						<textarea class="logs" cols="68" rows="6" name="inp{$useropt.id}" style="max-width:500px; max-height:250px">{$useropt.option_value|escape}</textarea>
-					{/if}
-					{if $useropt.option_input=="checkbox"}
-						<input type="checkbox" class="checkbox" name="inp{$useropt.id}" value="1" {if $useropt.option_value=="1"}checked="checked"{/if} />
-					{/if}		
+					<input type="checkbox" name="usePMR" class="public-setting" value="1" {if $usePMR==true}checked="checked"{/if} id="l_using_pmr" class="checkbox" />
+					<label for="l_using_pmr" class="public-setting">{t}I'm taking an PMR walkie talkie on channel 2 with me.{/t}</label>
+					<br />
 				{else}
-					{if $useropt.option_input=="checkbox"}
-						{if $useropt.option_value=="1"}
-							{t}Yes{/t}
-						{else}
-							{t}No{/t}
-						{/if}
+					{if $usePMR==true}
+						<span class="public-setting">{t}I'm taking an PMR walkie talkie on channel 2 with me.{/t}</span><br />
 					{else}
-						{$useropt.option_value|escape}
-					{/if}
-				{/if}
-			</td>
-			<td style="vertical-align:top;">
-				{if $edit==true}
-					{if $useropt.internal_use!=1}
-						<input type="checkbox" name="chk{$useropt.id}" value="1"{if $useropt.option_visible==1} checked="checked"{/if} class="checkbox" /> {t}show{/t}
-					{/if}
-				{else}
-					{if $useropt.internal_use!=1}
-						{if $useropt.option_visible==1}
-							<span style="color:#666666;">{t}visible{/t}</span>
-						{else}
-							<span style="color:#666666;">{t}invisible{/t}</span>
-						{/if}
-					{else}
-						<!-- <span style="color:#666666;">{t}internal{/t}</span> -->
+						{t}No. I'm not taking an PMR walkie talkie on channel 2 with me.{/t}<br /> <!-- TODO: Translation -->
 					{/if}
 				{/if}
 			</td>
 		</tr>
-	{foreachelse}
+		<tr><td class="spacer" colspan="2"></td></tr>
+
+		{if $errorUnknown==true}
+			<tr>
+				<td colspan="2">
+					<span class="errormsg">{t}An unknown error occured.{/t}</span>
+				</td>
+			</tr>
+		{/if}
+
 		<tr>
-			<td colspan="3">{t}No information on user details found.{/t}</td>
+			<td class="header-small" colspan="3">
+				{if $edit==true}
+					<input type="submit" name="cancel" value="{t}Cancel{/t}" class="formbutton" onclick="submitbutton('cancel')" />&nbsp;&nbsp;
+					<input type="submit" name="save" value="{t}Submit{/t}" class="formbutton" onclick="submitbutton('save')" />
+				{else}
+					<input type="submit" name="change" value="{t}Change{/t}" class="formbutton" onclick="flashbutton('change')" />
+				{/if}
+			</td>
 		</tr>
-	{/foreach}
-
-	<tr><td class="spacer" colspan="3"></td></tr>
-
-	<tr>
-		<td class="header-small" colspan="3">
-			{if $edit==true}
-				<input type="submit" name="cancel" value="{t}Cancel{/t}" class="formbutton" onclick="submitbutton('cancel')" />&nbsp;&nbsp;
-				<input type="submit" name="save" value="{t}Submit{/t}" class="formbutton" onclick="submitbutton('save')" />
-			{else}
-				<input type="submit" name="change" value="{t}Change{/t}" class="formbutton" onclick="flashbutton('change')" />
-			{/if}
-		</td>
-	</tr>
-	<tr><td class="spacer">&nbsp;</td></tr>
-</table>
-
+		<tr><td class="spacer">&nbsp;</td></tr>
+	</table>
 </form>
 
 {if $edit==false}
@@ -123,3 +100,6 @@
 		</table>
 	</form>
 {/if}
+
+
+
